Changed property @Autowired for constructor @Autowired.
Changed property @Autowired for constructor @Autowired:

The benefits of this is for instance, if you want to test the service
you don't need to depend on spring (or any other IoC) to inject the
dependent beans. By using contructor inection we can mock the dependent
beans and pass in as a normal contructor param.