oauth2 in rest1, web module acts as a client
Added oauth2 in rest 1 microservice.

Only grant types: authorization_code and client_credentials are
supported.

Web module acts as a client and connects to rest 1 service with
"client_credentials".
/Foos will get the results from Rest 1, using client_credentials (no
authentication from user required in web module)

/Courses will get the results from Rest 1, using client_credentials (
for this web endpoint the user has to be logged in in web module, then
for connecting to rest 1 service is the same as /foos, it will use
"client_credentials" -client_id & client_secret- to
authenticate/authorize with rest 1 service)